Add mac support
